feat(edit): enable ingredients to be removed when editing a recipe

+++ app/src/main/java/com/nicoapps/cooktime/data/dao/RecipeEntityDao.kt
@@ -28,5 +28,5 @@ interface RecipeEntityDao {
     fun getWithDetailsById(id: Long): Flow<RecipeWithDetailsEntity?>
 
     @Query("SELECT * FROM recipes WHERE id = :id")
-    fun getById(id: Int): Flow<RecipeEntity?>
+    fun getById(id: Long): Flow<RecipeEntity?>
 }
